simplify build now that git is baked into ruby-plus-devkit
With core/git not being fully capable of supporting a `bundle install`
because of semi-hardcoded paths to binaries, git is now baked into the
"devkit" portion of ruby-plus-devkit.

That means `bundle install` works, so the build reverts back to a
pattern established in the chef omnibus software definition:

1. bundle install the Ruby dependencies
2. gem install the local gems
3. install the git-ref'd gems

At the end of these steps, all the gems are "installed" in the standard
directories under GEM_HOME/gems and binstubs are in GEM_HOME/bin.

Appbundler is executed from GEM_HOME/bin to create our precisely-pinned
binstubs for only the gems that supply executables we expect users of
chef-client to use. Those appbundled binstubs are placed into the
package's bin directory under $pkg_prefix/bin.

With bundler working against, removed the clever gnarliness around parsing
the Gemfile.lock to install path- and git-based gems without bundler.