trim more fat and strip studio FS_ROOT from Gemfile
* Trimming the Fat

Exclude the byproducts of building gem extensions from the copy into
pkg_prefix.

* Stripping the FS_ROOT

The studio's FS_ROOT appears in the Gemfiles generated by appbundler.
Added a function to remove the build environment's FS_ROOT from a given
file. Currently only using it on the chef gem's Gemfile so that it
can be used with `bundle install` and `bundle exec` to test the build.